style: 운영진 리스트 페이지 및 상세 페이지 반응형 적용
