+++ src/components/molecules/OrganizerCards/index.module.scss
@@ -1,64 +1,80 @@
 @import '/src/styles/main';
 
-.projectWrapper {
+.organizersWrapper {
   display: grid;
   grid-template-columns: repeat(auto-fill, minmax(271px, 1fr));
-  gap: 34px 14px;
-  align-items: flex-start;
-  margin-bottom: 100px;
+  gap: 24px;
+  padding-bottom: 100px;
+  place-items: center;
 
   @include xs {
-    margin-bottom: 80px;
-    row-gap: 24px;
+    padding-bottom: 0;
   }
 
   .projectCardWrapper {
     display: flex;
     flex-direction: column;
-    gap: 8px;
     width: 100%;
+    max-width: 450px;
+    border-radius: 8px;
+    overflow: hidden;
+    box-shadow: 1px 1px 0 0 #404246 inset;
 
-    .thumbnail {
-      border-radius: 8px;
+    .emoji {
+      font-size: 220px;
+      line-height: 104%;
+      text-align: center;
+      height: 271px;
+      padding: 24px;
+      vertical-align: middle;
+    }
+
+    .profile {
       object-fit: cover;
       width: 100%;
       height: 100%;
-      aspect-ratio: 16 / 9;
+      aspect-ratio: 1 / 1;
     }
 
     .contents {
+      padding: 24px;
       width: 100%;
+      height: 160px;
       display: flex;
       flex-direction: column;
       gap: 8px;
+      background-color: color('white');
 
-      .flag {
+      .position {
         font-size: 12px;
         font-weight: 700;
         line-height: 18px;
-        color: color('info');
+        color: color('gray09');
       }
 
       .name {
         font-size: 20px;
         font-weight: 700;
         line-height: 28px;
-        color: color('gray01');
+        color: color('gray09');
       }
 
-      .title {
+      .introduction {
         font-size: 12px;
         font-weight: 400;
         line-height: 18px;
-        color: color('gray03');
+        color: color('gray08');
+        overflow: hidden;
+        text-overflow: ellipsis;
+        white-space: nowrap;
       }
 
-      .skillWrapper {
+      .stackWrapper {
         overflow: hidden;
         display: -webkit-box;
         -webkit-box-orient: vertical;
         -webkit-line-clamp: 1;
-        width: 271px;
+        width: 271px - 48px;
         color: color('gray04');
 
         :not(span:last-of-type) {
