+++ src/components/pages/OrganizerPage/index.module.scss
@@ -14,14 +14,6 @@
     h3 {
       @include text('H4/Bold', 'white');
     }
-
-    .samePositionOrganizers {
-      display: flex;
-      flex-flow: row wrap;
-      gap: 24px;
-      justify-content: flex-start;
-      align-items: flex-start;
-    }
   }
 
   .organizerMainContents {
@@ -30,7 +22,19 @@
     gap: 24px;
 
     @include sm {
+      gap: 32px;
       grid-template-columns: 1fr;
+    }
+
+    .divider {
+      display: none;
+
+      @include sm {
+        display: block;
+        width: 100%;
+        height: 1px;
+        background-color: color('gray04');
+      }
     }
 
     .organizerInfoWrapper {
@@ -40,11 +44,20 @@
       border-radius: 8px;
       box-shadow: 1px 1px 0 0 #404246 inset;
 
+      @include sm {
+        box-shadow: none;
+        gap: 24px;
+      }
+
       .thumbnailWrapper {
         display: flex;
         align-items: center;
         justify-content: center;
         padding: 24px;
+
+        @include sm {
+          padding: 24px 0;
+        }
 
         .emoji {
           font-size: 200px;
@@ -56,6 +69,16 @@
         .thumbnail {
           border-radius: 8px;
           object-fit: cover;
+          width: 100%;
+          height: 100%;
+          aspect-ratio: 1 / 1;
+          max-width: 224px;
+          max-height: 224px;
+
+          @include sm {
+            max-width: 400px;
+            max-height: 400px;
+          }
         }
       }
 
@@ -65,6 +88,11 @@
         flex-direction: column;
         align-items: flex-start;
         gap: 8px;
+
+        @include sm {
+          padding: 0;
+          gap: 16px;
+        }
 
         .position {
           @include text('body1/Bold', 'primary');
@@ -98,8 +126,17 @@
       border-radius: 8px;
       box-shadow: 1px 1px 0 0 #404246 inset;
 
+      @include sm {
+        box-shadow: none;
+        padding: 0;
+      }
+
       h1 {
         @include text('H4/Bold', 'gray01');
+
+        @include xs {
+          @include text('H5/Bold');
+        }
       }
 
       .organizerDescriptionItem {
@@ -119,7 +156,11 @@
           word-wrap: break-word;
           white-space: pre-line;
 
-          @include text('body2/Regular', 'gray03');
+          @include text('body1/Regular', 'gray03');
+
+          @include xs {
+            @include text('body2/Regular');
+          }
         }
 
         .applyButton {
