+++ src/lib/apis/organizer.ts
@@ -26,7 +26,7 @@ export function getOrganizer({ id }: { id: number; }) {
 export function getOrganizerCount() {
   const organizers = organizersData as Organizer[];
 
-  return organizers.filter((organizer) => !organizer.isArchived && organizer.dndPosition !== '마스코트').reduce(
+  return organizers.filter((organizer) => !organizer.isArchived).reduce(
     (acc, { dndPosition }) => ({
       ...acc,
       [dndPosition]: checkNumber(acc[dndPosition]) + 1,
